Finished refactoring test cases
